requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pymupdf==1.26.4
pillow==10.4.0
pytesseract==0.3.13
transformers==4.44.2
torch==2.4.1
huggingface-hub==0.24.6
